requirements.txt
fastapi
uvicorn
pandas
python-docx
PyPDF2
sentence-transformers
torch
